feat: move contents of eventTypeFacade to AgencyFacade and add full meeting slug to teamEventType
